refactor: request,responseDTO를 core모듈로 이동
